packages/prompter/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comfyui-prompter"
version = "0.1.0"
description = "GUI and REST API for ComfyUI workflow recommendation and generation"
readme = "../../README.md"
requires-python = ">=3.10"
license = "Apache-2.0"

dependencies = [
    "comfyui-agent-sdk>=0.1.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "ollama>=0.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]

[project.scripts]
comfyui-gui = "main:main"
comfyui-api = "api_server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
